refactor: Service 인터페이스들을 domain 패키지로 이동
