+++ src/main/java/com/example/momo/domain/meeting/application/MeetingServiceImpl.java
@@ -6,9 +6,14 @@ import static com.example.momo.global.rabbitmq.constant.RoutingKeys.*;
 import java.time.LocalDateTime;
 import java.util.List;
 
+import com.example.momo.domain.meeting.domain.MeetingOutboxService;
+import com.example.momo.domain.meeting.domain.MeetingPaymentOutboxService;
+import com.example.momo.domain.meeting.domain.MeetingService;
 import com.example.momo.global.rabbitmq.dto.common.EventWrapper;
 import com.fasterxml.jackson.core.JsonProcessingException;
+
 import lombok.extern.slf4j.Slf4j;
+
 import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.context.ApplicationEventPublisher;
@@ -51,8 +56,6 @@ import com.example.momo.global.webclient.user.dto.UserClientResponseDto;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 import lombok.RequiredArgsConstructor;
-
-
 
 @Slf4j
 @Service
@@ -154,7 +157,7 @@ public class MeetingServiceImpl implements MeetingService {
 
 	@Override
 	@Transactional(readOnly = true)
-	@Cacheable(value = "meeting", key = "#meetingId",cacheManager = "cacheManager")
+	@Cacheable(value = "meeting", key = "#meetingId", cacheManager = "cacheManager")
 	public MeetingResponseDto getMeeting(Long meetingId) {
 
 		Meeting meeting = meetingRepository.findById(meetingId).orElseThrow(() ->
@@ -332,18 +335,18 @@ public class MeetingServiceImpl implements MeetingService {
 			throw new MeetingException(MeetingExceptionCode.INSUFFICIENT_SCORE);
 		}
 
-			// 참가자 수 확인
-			if (meeting.getCurrentParticipantsCount() >= meeting.getMaxParticipantsCount()) {
-				throw new MeetingException(MeetingExceptionCode.MEETING_IS_FULL);
-			}
-			MeetingParticipant participant = MeetingParticipant.createParticipant(meeting, userId);
-			meeting.addMeetingParticipant(participant);
+		// 참가자 수 확인
+		if (meeting.getCurrentParticipantsCount() >= meeting.getMaxParticipantsCount()) {
+			throw new MeetingException(MeetingExceptionCode.MEETING_IS_FULL);
+		}
+		MeetingParticipant participant = MeetingParticipant.createParticipant(meeting, userId);
+		meeting.addMeetingParticipant(participant);
 
-			MeetingElasticsearchOutbox esOutbox = new MeetingElasticsearchOutbox(meeting.getId(),
-				ElasticsearchEventType.SAVE);
-			meetingOutboxService.saveMeetingOutbox(esOutbox);
+		MeetingElasticsearchOutbox esOutbox = new MeetingElasticsearchOutbox(meeting.getId(),
+			ElasticsearchEventType.SAVE);
+		meetingOutboxService.saveMeetingOutbox(esOutbox);
 
-			eventPublisher.publishEvent(new MeetingElasticEvents.Save(meeting, esOutbox.getId()));
+		eventPublisher.publishEvent(new MeetingElasticEvents.Save(meeting, esOutbox.getId()));
 
 		// 참가비 무료일 경우 즉시 참가자 추가
 		if (meeting.getParticipationFee() == 0) {
