+++ src/main/java/com/example/momo/domain/meeting/event/springEvents/MeetingEventListener.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Component;
 import org.springframework.transaction.event.TransactionPhase;
 import org.springframework.transaction.event.TransactionalEventListener;
 
-import com.example.momo.domain.meeting.application.MeetingOutboxService;
+import com.example.momo.domain.meeting.domain.MeetingOutboxService;
 import com.example.momo.domain.meeting.domain.MeetingRepository;
 
 import lombok.RequiredArgsConstructor;
