+++ src/main/java/com/example/momo/domain/payment/application/scheduler/PaymentOutboxScheduler.java
@@ -9,7 +9,7 @@ import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 
-import com.example.momo.domain.payment.application.PaymentOutboxService;
+import com.example.momo.domain.payment.domain.PaymentOutboxService;
 import com.example.momo.domain.payment.event.rabbitmq.producer.PaymentEventProducer;
 
 import lombok.extern.slf4j.Slf4j;
