+++ src/main/java/com/example/momo/domain/user/application/UserOutboxCleanupScheduler.java
@@ -6,6 +6,7 @@ import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 
 import com.example.momo.domain.user.domain.UserOutboxEvent;
+import com.example.momo.domain.user.domain.UserOutboxService;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
