+++ src/main/java/com/example/momo/domain/user/application/UserServiceImpl.java
@@ -26,8 +26,10 @@ import com.example.momo.domain.user.application.dto.UserResponseDto;
 import com.example.momo.domain.user.application.dto.WithdrawRequestDto;
 import com.example.momo.domain.user.domain.User;
 import com.example.momo.domain.user.domain.UserCategory;
+import com.example.momo.domain.user.domain.UserOutboxService;
 import com.example.momo.domain.user.domain.UserRating;
 import com.example.momo.domain.user.domain.UserRepository;
+import com.example.momo.domain.user.domain.UserService;
 import com.example.momo.domain.user.event.springEvent.UserEvents;
 import com.example.momo.domain.user.exception.UserErrorCode;
 import com.example.momo.domain.user.exception.UserException;
